tests, network: Extract macvtap tests to a dedicated file
